# bootstrap/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Cerebro API
  description: |
    Contrato único dos endpoints HTTP expostos pela aplicação **até ao momento**.

    - Rotas **Apache Camel** (servlet em `/api/*`): prefixo `/api/v1/...`.
    - Rotas **Spring MVC** (DispatcherServlet na raiz): prefixo `/v1/...` (sem `/api`).

    Autenticação: **Firebase ID Token** (Bearer) para dashboard, analytics e sessão; `POST /v1/auth/register` exige conta Firebase ainda sem tenant (convite).
    Outras rotas Camel permanecem abertas salvo indicação em contrário.
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
...
paths:
...
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Associar conta Firebase a um tenant (convite)
      description: |
        Requer `Authorization: Bearer` com ID Token Firebase de utilizador **ainda sem** linha em `portal_user`.
        O corpo contém o código de convite em texto; o backend armazena apenas o hash SHA-256.
      operationId: postPortalRegister
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortalRegisterRequest"
      responses:
        "200":
          description: Tenant associado e perfil inicial (BASIC)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalRegisterResponse"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestErrorResponse"
        "401":
          description: Token em falta ou inválido
        "403":
          description: Conta já associada a um tenant
        "409":
          description: Utilizador já registado ou convite inválido
  /v1/auth/me:
    get:
      tags: [Auth]
      summary: Dados da sessão (ID Token Firebase)
      operationId: getPortalSession
      security:
        - bearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalSessionResponse"
        "401":
          description: Token em falta ou inválido
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID Token (claims `tenant_id` / `profile_level` após registo)
  schemas:
    PortalRegisterRequest:
      type: object
      required: [inviteCode]
      properties:
        inviteCode:
          type: string
          description: Código de convite em texto claro
    PortalRegisterResponse:
      type: object
      required: [tenantId, profileLevel]
      properties:
        tenantId:
          type: string
        profileLevel:
          type: string
          enum: [BASIC, PRO, ULTRA]
    PortalSessionResponse:
      type: object
      required: [tenantId, profileLevel]
      properties:
        tenantId:
          type: string
        profileLevel:
          type: string
          enum: [BASIC, PRO, ULTRA]
...
    IngestErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...